make the main activity of qr coder merely serve as a director for other activities.
functionally this will make it seem like the user is by default scanning whenever they start the qr module.

this is a little more complicated that just directly using the CaptureActivity as the main activity, but since the CaptureActivity
is a third party code, this requires less modification of third party code.